requirements.txt
feedparser>=6.0.0
requests>=2.28.0
python-dotenv>=0.19.0
tushare>=1.2.0
pandas>=1.0.0
